renaming variable overview to project-overview
