+++ src/app/components/navigation/navigation.html
@@ -25,7 +25,7 @@
       <md-icon md-font-icon="fa fa-table fa-fw"></md-icon>
       <p>Plate Viewer</p>
     </md-list-item>
-    <md-list-item ng-click="$ctrl.$state.go('overview')">
+    <md-list-item ng-click="$ctrl.$state.go('project-overview')">
       <md-icon md-font-icon="fa fa-desktop fa-fw"></md-icon>
       <p>Project Overview</p>
     </md-list-item>
